Add word entry

+++ src/main/java/com/example/VerboQuest/repository/WordRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WordRepository extends CrudRepository<Word, Integer> {
+public interface WordRepository extends CrudRepository<Word, Long> {
 }
